Renamed as `manager`
Renamed as `manager` as it makes sense what they does also they ends with `Manager`
