+++ V2rayNG/app/src/main/kotlin/com/v2ray/ang/fmt/TrojanFmt.kt
@@ -5,7 +5,7 @@ import com.v2ray.ang.dto.EConfigType
 import com.v2ray.ang.dto.ProfileItem
 import com.v2ray.ang.dto.V2rayConfig.OutboundBean
 import com.v2ray.ang.extension.idnHost
-import com.v2ray.ang.handler.MmkvManager
+import com.v2ray.ang.manager.MmkvManager
 import com.v2ray.ang.util.Utils
 import java.net.URI
 import kotlin.text.orEmpty
