+++ V2rayNG/app/src/main/kotlin/com/v2ray/ang/manager/MigrateManager.kt
@@ -1,4 +1,4 @@
-package com.v2ray.ang.handler
+package com.v2ray.ang.manager
 
 import android.util.Log
 import com.tencent.mmkv.MMKV
@@ -7,7 +7,7 @@ import com.v2ray.ang.AppConfig.ANG_PACKAGE
 import com.v2ray.ang.dto.EConfigType
 import com.v2ray.ang.dto.ProfileItem
 import com.v2ray.ang.dto.ServerConfig
-import com.v2ray.ang.handler.MmkvManager.decodeServerConfig
+import com.v2ray.ang.manager.MmkvManager.decodeServerConfig
 import com.v2ray.ang.util.JsonUtil
 import com.v2ray.ang.util.Utils
