+++ V2rayNG/app/src/main/kotlin/com/v2ray/ang/manager/MmkvManager.kt
@@ -1,4 +1,4 @@
-package com.v2ray.ang.handler
+package com.v2ray.ang.manager
 
 
 import com.tencent.mmkv.MMKV
