+++ V2rayNG/app/src/main/kotlin/com/v2ray/ang/receiver/BootReceiver.kt
@@ -3,7 +3,7 @@ package com.v2ray.ang.receiver
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import com.v2ray.ang.handler.MmkvManager
+import com.v2ray.ang.manager.MmkvManager
 import com.v2ray.ang.service.V2RayServiceManager
 
 class BootReceiver : BroadcastReceiver() {
