+++ V2rayNG/app/src/main/kotlin/com/v2ray/ang/ui/SubSettingActivity.kt
@@ -13,8 +13,8 @@ import com.v2ray.ang.databinding.ActivitySubSettingBinding
 import com.v2ray.ang.databinding.LayoutProgressBinding
 import com.v2ray.ang.dto.SubscriptionItem
 import com.v2ray.ang.extension.toast
-import com.v2ray.ang.handler.AngConfigManager
-import com.v2ray.ang.handler.MmkvManager
+import com.v2ray.ang.manager.AngConfigManager
+import com.v2ray.ang.manager.MmkvManager
 import com.v2ray.ang.helper.SimpleItemTouchHelperCallback
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
