+++ V2rayNG/app/src/main/kotlin/com/v2ray/ang/util/Utils.kt
@@ -24,7 +24,7 @@ import com.v2ray.ang.BuildConfig
 import com.v2ray.ang.R
 import com.v2ray.ang.dto.Language
 import com.v2ray.ang.extension.toast
-import com.v2ray.ang.handler.MmkvManager
+import com.v2ray.ang.manager.MmkvManager
 import com.v2ray.ang.service.V2RayServiceManager
 import java.io.IOException
 import java.net.*
